tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.stroke.skew.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.stroke.skew</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.stroke.skew</h1>
<p class="desc">Strokes lines are skewed by the current transformation matrix</p>


<script>
var t = async_test("Strokes lines are skewed by the current transformation matrix");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);

  ctx.save();
  ctx.beginPath();
  ctx.moveTo(49, -50);
  ctx.lineTo(201, -50);
  ctx.rotate(Math.PI/4);
  ctx.scale(1, 283);
  ctx.strokeStyle = '#0f0';
  ctx.stroke();
  ctx.restore();

  ctx.save();
  ctx.beginPath();
  ctx.translate(-150, 0);
  ctx.moveTo(49, -50);
  ctx.lineTo(199, -50);
  ctx.rotate(Math.PI/4);
  ctx.scale(1, 142);
  ctx.strokeStyle = '#f00';
  ctx.stroke();
  ctx.restore();

  ctx.save();
  ctx.beginPath();
  ctx.translate(-150, 0);
  ctx.moveTo(49, -50);
  ctx.lineTo(199, -50);
  ctx.rotate(Math.PI/4);
  ctx.scale(1, 142);
  ctx.strokeStyle = '#f00';
  ctx.stroke();
  ctx.restore();

  _assertPixel(canvas, 0,0, 0,255,0,255);
  _assertPixel(canvas, 50,0, 0,255,0,255);
  _assertPixel(canvas, 99,0, 0,255,0,255);
  _assertPixel(canvas, 0,25, 0,255,0,255);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 99,25, 0,255,0,255);
  _assertPixel(canvas, 0,49, 0,255,0,255);
  _assertPixel(canvas, 50,49, 0,255,0,255);
  _assertPixel(canvas, 99,49, 0,255,0,255);
  t.done();

});
</script>
